Separated out provided request tests into separate files.
